Small fixes.

+++ frontend/localbusiness/src/components/Partials/RemoveFavorite.js
@@ -1,39 +1,37 @@
-import React from 'react';
-import {read_cookie} from "sfcookies";
+import React, {useContext} from 'react';
+import {BaseContext} from "../ContextProviders/BaseContextProvider";
 
 const RemoveFavorite = (props)=>{
 
-    // console.log('Remove props ', props)
+    const [base] = useContext(BaseContext);
 
     const removeFavorite = ()=>{
-        //console.log("removeFavorite...")
 
-        const cookie = read_cookie('credentials');
-        const userId = cookie.id;
-        const businessId = props.businessId;
-        const accessToken = cookie.access_token;
-        const url =  `http://localhost/public/api/favorites/delete`;
+        if(base){
+            let myHeaders = new Headers();
+            myHeaders.append("Authorization", `Bearer ${base.credentials.accessToken}`);
 
-        var myHeaders = new Headers();
-        myHeaders.append("Authorization", `Bearer ${accessToken}`);
+            let urlencoded = new URLSearchParams();
+            urlencoded.append("user_id", base.credentials.userId);
+            urlencoded.append("business_id", props.businessId);
 
-        var urlencoded = new URLSearchParams();
-        urlencoded.append("user_id", userId);
-        urlencoded.append("business_id", businessId);
+            let requestOptions = {
+                method: 'DELETE',
+                headers: myHeaders,
+                body: urlencoded,
+                redirect: 'follow'
+            };
 
-        var requestOptions = {
-            method: 'DELETE',
-            headers: myHeaders,
-            body: urlencoded,
-            redirect: 'follow'
-        };
+            fetch(base.urls.favoritesDelete, requestOptions)
+                .then(response => response.json())
+                .then(data => setMyStates(data))
+                .catch(error => console.log('error', error));
+        }
 
-        fetch(url, requestOptions)
-            .then(response => response.text())
-            .then(result => console.log(result))
-            .then(props.funcRefs(false))
-            .catch(error => console.log('error', error));
+    }
 
+    const setMyStates = (data)=>{
+        props.funcRefs(false);
     }
 
     return <button href="#" onClick={removeFavorite} className="btn btn-danger favorito"><em className="fa fa-heart"></em> Remover Favorito </button>
